Created entities folder for test cases and refactored games_manager to use_case_implementations and data to data_gateways
