+++ src/test/java/entities/CellTest.java
@@ -1,0 +1,4 @@
+package entities;
+
+public class CellTest {
+}
