+++ src/test/java/use_case_implementations/EndGameManagerTest.java
@@ -1,4 +1,4 @@
-package games_manager;
+package use_case_implementations;
 
 import data_gateways.GameCreator;
 import entities.Game;
