+++ src/test/java/use_case_implementations/PlayerManagerTest.java
@@ -1,4 +1,4 @@
-package games_manager;
+package use_case_implementations;
 
 import usecases.usecase_implementations.PlayerManager;
 import entities.*;
